test: convert integration to unit test
